refactor:메인 로고 애니메이션 수정
