+++ src/components/main/HowText.tsx
