+++ src/components/main/SectionMain.tsx
@@ -9,7 +9,7 @@ const SectionMain: React.FC<SectionProps> = ({ className }) => {
   return (
     <>
       <div
-        className={` flex flex-col justify-start  items-center mt-[-120px] ${className}`}
+        className={`flex flex-col justify-start items-center mt-[-120px] ${className}`}
       >
         <TextFade direction={'up'} duration={2.2} delay={0.2}>
           <div className="font-staatliches text-[clamp(3rem,20vw,13rem)]">
